% File src/R/man/NenS-package.Rd
% $Id: delftfews-package.Rd 115 2011-09-19 10:09:34Z mariotomo $

\name{delftfews-package}
\Rdversion{1.1}
\alias{delftfews-package}
\alias{delftfews}
\alias{[.zoo}
\alias{[<-.zoo}
\alias{$<-.zoo}
\alias{$.zoo}
\alias{Ops.zoo}
\docType{package}
\title{
alfa version of delftfews
}
\description{
  delftfews (mostly) defines time series and PI files.

  time series in delftfews are not meant for statistical analysis but
  for simpler operations where the data.frame shape is more in order.
}
\details{
  this package comes from the joint work of Delft based Deltares and
  Utrecht private company Nelen & Schuurmans.
}
\author{
   Mario Frasca <mario.frasca@nelen-schuurmans.nl>
   Mich\`el van Leeuwen <michel.vanleeuwen@nelen-schuurmans.nl>
}
\keyword{ package }
